switched run details to component
