+++ src/client/app/templates/index.hbs
@@ -38,7 +38,8 @@
             {{#if runInstance.isrunning}}
               <td>running</td>
             {{else}}
-              <td>{{#link-to "editRuleset.run" "_run_" runInstance.id (query-params run=true)}}Details{{/link-to}}</td>
+              <td>{{#link-to "run" runInstance.id}}Details {{/link-to}}</td>
+
             {{/if}}
 
           </tr>
