+++ src/client/app/templates/run.hbs
@@ -1,0 +1,11 @@
+<div style="float:none;display:flex;">
+	{{#link-to 'index' tagName='button' class='btn btn-success btn-sm btn-return-title'}}
+		{{#bs-tooltip placement="left"}}Return to main page{{/bs-tooltip}}
+        <i class="fa fa-angle-double-left"></i>
+	{{/link-to}}
+    <h2>Run Details</h2>
+</div>
+<div>
+	{{run-details model=model type=type page=page ruleid=ruleid}}
+</div>
+{{outlet}}
